pyproject.toml
[project]
name = "intern-rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "fastapi>=0.121.0",
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.4",
    "langchain-groq>=1.0.0",
    "langchain-pinecone>=0.2.13",
    "pinecone-client>=6.0.0",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "uvicorn>=0.38.0",
]
